bfd/mrk3: Change Verilog DAT format for MRK3.
The verilog DAT format required for MRK3 is different (slightly) to the
format currently supported in bfd.

This commit modifies the existing code to support the format desired for
mrk3.  Given that I've been unable to find an "official" specification
for the verilog DAT format it is unlikely that these changes will be
pushed back upstream.

bfd/ChangeLog:

	* verilog.c (MRK3_VERILOG_HACKS): Define.
	(verilog_write_address): Handle MRK3 specifc 32-bit addresses.
	(verilog_write_section): Handle MRK3 specific file layout.
	(verilog_scan): Merge adjacent lines for MRK3.
	(verilog_read_section): Handle MRK3 specific file layout.