gdb: Check the selected-frame in frame_find_by_id.
In gdb we track two different things, the current-frame, and the
selected-frame.  The current-frame represents the frame in which we are
currently stopped.  If gdb stops at a breakpoint and the user asks for a
backtrace then all the frames in the backtrace will be the ancestors of
the the current-frame.

Also the current-frame does not change until the inferior is set running
again (though it is possible for the user to alter the contents of the
current-frame this is not the same as changing the current-frame).

There is also the selected-frame.  This is the frame that the user has
"selected" using the frame selection commands 'frame', 'up', or 'down'
being the most common.

Usually the selected-frame is always an ancestor of the current-frame,
however, there is a feature of the 'frame' command that allows arbitrary
frames to be created, in this case the selected-frame might not be an
ancestor of the current-frame.

The problem this causes is that lazy register values hold the frame-id
for the frame in which the real register value can be obtained.  When
gdb needs to fetch the actual register value we try to find the frame
using frame_find_by_id, which currently looks at the current-frame, and
all the ancestors of the current-frame.

When the selected-frame is not an ancestor of the current-frame then
frame_find_by_id will fail to find the required frame, and gdb will fail
with an assertion while trying to fetch the lazy value (the assertion is
that the frame will always be found).

This patch extends frame_find_by_id to also check the selected-frame as
a separate comparison as well as the current-frame and its ancestors.
This resolves the issue.

There are also two new tests added that show this issue.  The first is
an mi test using var objects, which is where I first saw this issue.
The second is a slightly simpler test, just creating a new frame then
calling 'info frame' is enough to trigger the assertion.

gdb/ChangeLog:

	* frame.c (frame_find_by_id): Also check the selected frame.

gdb/testsuite/ChangeLog:

	* gdb.mi/mi-var-frame.c: New file.
	* gdb.mi/mi-var-frame.exp: New file.
	* gdb.base/create-frame.c: New file.
	* gdb.base/create-frame.exp: New file.