gdb: Check the selected frame in frame_find_by_id.
It is possible that the selected frame is not an ancestor of the current
frame.  This can happen when the frame command is use to create a new
frame by specifying an address.

When resolving lazy values we rely on frame_find_by_id finding the frame
in which the value was created.  Currently this is done by looking at
the current frame and all ancestors, however, if the selected frame is
not an ancestors then this is not good enough.

This commit extends frame_find_by_id to also check the selected frame,
if non of the ancestors of the current frame match.  In most cases the
current frame is also an ancestors, so this might result in duplicate
checks, however, this penalty should be acceptably small.

gdb/ChangeLog:

	* frame.c (frame_find_by_id): Also check the selected frame.