gdb: Select a frame for frame_info.
Within the 'info frame' command (implemented in the frame_info function)
gdb will create register values, which are then displayed to the user.
These register values will be created as lazy value.  As part of
fetching the value of a lazy register value we must find the frame in
which the value was created based on the frame id, which is stored in
the value, this is done with a call to frame_find_by_id.

In a previous commit I already addressed the problem that
frame_find_by_id did not used to consider the selected frame when trying
to find a frame.  This is now fixed.

However, if in frame_info the user passed a frame specification, then it
is possible that a brand new frame will have been created, a frame
outside of the current backtrace, and also not the currently selected
frame.  So, usually, something like, 'info frame 4' will cause gdb to
crash with an assertion failure.  This can be especially annoying if the
user accidentally types 'info frame 4' intending to ask about the fourth
stack frame, but, there happens to only be 3 stack frames at present.

The solution I propose here is that frame_info should temporarily select
the frame identified by the user.  The previously selected frame will be
restored at the end of the function.

gdb/ChangeLog:

	* frame.h (make_restore_selected_frame_cleanup): Declare new
	function.
	* frame.c (do_restore_selected_frame): New function.
	(make_restore_selected_frame_cleanup): Define new function.
	* stack.c (frame_info): Temporarily select frame for duration of
	this function.

gdb/testsuite/ChangeLog:

	* gdb.base/create-frame.exp: Add test for 'info frame' with
	specific address.