bfd: Ignore sym value when patching relocs against section syms.
When patching relocations the value patched into the relocation
is (initially at least) calculated as the base address of the section
containing the symbol (being relocated against) and the value of the
symbol.

There is an issue however, when the symbol in question is a section
symbol, a symbol that points to the start of a section.

After performing a non-relocatable link the value of a section symbol is
adjusted to be the virtual address of the section.  This causes problems
if we then re-link the executable, as the initial value is now computed
as the base address of the section plus the symbol value, which is also
the base address of the section, with the result that the address being
patched in is now wrong.

This commit attempts to work around this issue by tweaking the code that
computes the initial value that will be patched in.  For section symbols
we now use solely the virtual address of the section.

bfd/ChangeLog:

	* elf64-mrk3.c (mrk3_elf_relocate_section): For section symbols
	use the section base address, and discard the symbol value.