objcopy/strip: Add option for remove-relocations.
When objcopy and strip are used these tools use the bfd library to load
the input file, modifications are made, and the output file is then
written out.

Relocations are held within a section of the input file, and a user
might think that relocations can be removed by using the command line
option `remote-section'.  However, this does not work; when the input
file is loaded the relocation sections are not managed in the same way
as other sections, instead they attach themselves to the section for
which the relocations apply, and are held not as a section, but as a
list of relocations.  Within the objcopy and strip tools there is a
specific function which is responsible for copying the relocations from
the input file to the output file.

This commit adds new command line options for the purpose of removing
relocations.  The new option takes the name of a section and removes the
relocations for that section from the output file.

As an example, the following shows how objcopy can be used to remove the
relocations for section '.text' from file 'test.exe'.

    objcopy --remove-relocations=.text test.exe

binutils/ChangeLog:

	* doc/binutils.texi (objcopy): Document 'remove-relocations'.
	(strip): Likewise.
	* objcopy.c (SECTION_CONTEXT_REMOVE_RELOCS): Define.
	(enum command_line_switch): Add 'OPTION_REMOVE_RELOCS'.
	(struct option strip_options): Add 'remove-relocations'.
	(struct option copy_options): Likewise.
	(copy_usage): Likewise.
	(strip_usage): Likewise.
	(handle_remove_relocations_option): New function.
	(discard_relocations): New function.
	(handle_remove_section_option): New function.
	(copy_relocations_in_section): Use discard_relocations.
	(strip_main): Use handle_remove_section_option for
	'remove-section', and handle 'remove-relocations' option.
	(copy_main): Likewise.