mrk3/bfd: Adjust relocation addend after patching relocation.
Commit 04562a8137e981597edfd637e09f94b029f7d93f extended the addend
adjustment, which previously only happened when performing a relocatable
link, to also happen in the emit-relocations case.   The problem is that
this adjustment was happening too early, before the relocation was being
patched, which caused the relocation to fail to apply in some cases.

This restructuring of the code moves the patching of the relocation to
before the addend is adjusted, this should now be correct.

bfd/ChangeLog:

	* elf64-mrk3.c (mrk3_elf_relocate_section): Move the addend
	adjustment to after the relocations have been patched.