mrk3: Dump contents of .mrk3.records section.
Add support for dumping the contents of the .mrk3.records section.  This
section is used to hold information about ".align" and ".org" directives
in the assembler code.  This information will (in a future commit) be
used to ensure that ".align" and ".org" directives are honoured even
when linker relaxation is used.

bfd/ChangeLog:

	* elf64-mrk3.h: New file.
	* elf64-mrk3.c: Add elf64-mrk3.h include.
	(get_elf_r_symndx_section): New function.
	(get_elf_r_symndx_offset): New function.
	(internal_reloc_compare): New function.
	(mrk3_is_section_for_address): New function.
	(struct mrk3_find_section_data): New structure.
	(mrk3_find_section_for_address): New function.
	(+elf64_mrk3_load_records_from_section): New function.
	(elf64_mrk3_load_property_records): New function.
	(elf64_mrk3_property_record_name): New function.

binutils/ChangeLog:

	* configure.ac: Setup private mrk3 objdump functions and files.
	* configure: Regenerate.
	* objdump.h (objdump_private_desc_elf64_mrk3): New declaration.
	* od-mrk3.c: New file.