ld: Add new LOAD section type.
This commit introduces a new section type LOAD, this is the opposite
of the NOLOAD section type.  An output section marked as LOAD will
always be flagged as loadable, and given that being loadable without
being allocatable does not make sense, the section will also be
allocatable.

This section type can be used to force an output section containing
bss sections to be a loadable section.  The content of the bss sections
will become all 0's (zeros), after which the bss output sections will be
no different to standard data sections.

ld/ChangeLog:

	* ld.texinfo (Output Section Type): Add information about LOAD
	section type.
	* ldgram.y (type): Add LOAD type.
	* ldlang.c (lang_add_section): Handle isload_section.
	(map_input_to_output_sections): Likewise.
	* ldlang.h (enum section_type): Add isload_section type.
	* ldlex.l ("LOAD"): Make available for linker scripts.