bfd: Linker relaxation, fix reloc/symbol adjustment.
When adjusting the relocations and symbols during linker relaxation, the
wrong address range was used if an alignment directive appeared later
within the section than the bytes being deleted.

Consider the following example, showing a small region of a section,
from which 4 bytes are being deleted.  There's an alignment point in the
section too:

            Delete 4 bytes from here.
            |
    -----------------------------------------------
    | | | | | | | | | | | | | | | | | | | | | | | |
    -----------------------------------------------
            \_______/\_____________/|
              Delete      Move      Alignment directive here.

            \_______________/---- Previously, incorrect region checked
                                  for relocations.

            \______________________/---- Updated region checked for
                                         relocations.

                            \______/---- Bug range.

After deleting the bytes shown, the bytes labelled "Move" are moved down
(to a lower address) so that they start from the point of the deletion.
We then check for any relocations that also need to move, this is where
the bug was.  The diagram above shows the region that was checked before
this commit, and the region that is checked after this commit.  Any
relocations that (before this commit) were in the marked "Bug range"
would not be moved correctly; the relocation would remain in the wrong
place.

bfd/ChangeLog:

	* elf64-mrk3.c (mrk3_elf_relax_delete_bytes): Remove adjustment of
	toaddr.