bfd/ld: Update .mrk3.records sections during link.
The .mrk3.records sections contain a version number as the first 2-byte
value in the section.  When a partial link is performed then many
.mrk3.records sections are joined together into a single .mrk3.records
output section.

However, if the contents of each .mrk3.records input section are simply
concatenated together then the final result will be incorrect, it will
have the version number repeated throughout the contents of the output
section.

The solution is to remove the version number from all but the first
input section before we merge the input sections together into the
output section, in this way, the produced output section will be valid,
it has a single version number at the beginning, followed by all of the
records.

bfd/ChangeLog:

	* elf64-mrk3.h (elf64_mrk3_update_records_section): Declare.
	* elf64-mrk3.c (elf64_mrk3_update_records_section): New function.

ld/ChangeLog:

	* emultempl/mrk3.em: Add 'elf64-mrk3.h' include.
	(mrk3_after_allocation): New function.