bfd/mrk3: Fix relocation addend adjustment in linker relaxation.
Consider the following situation, a section in which we want to delete
some bytes, shown as 'XX'. In this section is a symbol 'S', and a
relocation 'R' targets the symbol 'S', but also has an addend that
places the actual target of the relocation beyond the deleted bytes.

As an added complication, the section also contains an alignment point,
shown as 'A'.  When we delete the bytes 'XX' only the section content
between the 'XX' and the 'A' will be moved, in order to preserve the
alignment requirement, the 'A' will not move.

    Relocation (R)
      |
      |    (addend of R)
      |--------------------.
      |                    |
      v                    v

    |-S-----XX--------A--------------|     <-- A section

      ^     ^          ^
      |     |          \---- Alignment point.
      |     |
      |     \---- Some bytes to delete.
      |
      \---- A symbol.

Now hopefully it should be clear that the addend of 'R' should not be
adjusted in this case, the destination of the relocation has not moved.
However, if the addend was such that the destination of the relocation
was between the 'XX' and the 'A' then the addend would need to be
adjusted.

Before the introduction of the alignment point tracking the case shown
above could not occur, in all cases the addend would need to be
adjusted.  However, since the introduction of the alignment point
tracking, we do need consider the above case.

This commit adds support for the above case, and should ensure that
relocation addends are not incorrectly decremented when they should not
be.

bfd/ChangeLog:

	* elf64-mrk3.c (mrk3_elf_relax_delete_bytes): Fix calculation of
	shrinked_insn_address.  When updating relocation addends, don't
	adjust relocations when the addend spans into a region of code
	that was not moved.