mrk3/ld: Report expression location for data statements errors.
If there's an error in a data statement within a linker script, for
example, an undefined symbol, then we should report the location, within
the linker script of the error.  Currently, we just report that an error
occurred.

For example, in a linker script test.ld:

  SECTIONS
  {
    .data :
    {
      LONG (__undefined_symbol);
    }
  }

Currently the linker reports:

  ld: invalid data statement

With this commit the linker reports:

  test.ld:5: invalid data statement

This is a small improvement.  Later commits will improve on the error
reporting event further by making use of more specific error messages,
however, this "invalid data statement" error is the error of last
resort, and so it's worth having the file and line number included in
this error.

ld/ChangeLog:

	* ldlang.c (lang_do_assignments_1): Report expression location for
	linker errors from data statements.