mrk3/gdb: Restructure of prologue analysis code.
This commit unifies the two places where we analyse function prologues
into a single function mrk3_analyze_prologue.  The inspiration for this
code structure was taken from the arm-tdep.c code.

The unification has forced some code that was in mrk3_analyze_prologue
to move out into mrk3_frame_cache.  While doing this I've managed to
simplify things like mrk3_frame_prev_register; removing the special case
handling of the pc register, this is all handled in mrk3_frame_cache
now.

I have also introduced GDB to the idea of CALL vs ECALL functions, for
which we now have basic support.  Specifically, GDB will now try to find
the end of the function, and look to see if the function ends in a CALL
or ECALL instruction.  This will not catch all cases, and there is more
that we can do to improve the reliability of function type analysis, but
these improvements can come later.

gdb/ChangeLog:

	* mrk3-tdep.c (struct mrk3_frame_cache): Delete 'base', add
	'frame_size', 'frame_base', and 'frame_type' members.
	(mrk3_analyze_prologue): Forward declare.  Rewrite to not rely on
	'struct frame_info' so that the code can be used from
	'mrk3_skip_prologue'.  Functionality to load actual values from
	the frame is moved out to 'mrk3_frame_cache'.
	(mrk3_skip_prologue): Rewrite to use mrk3_analyze_prologue.
	(mrk3_analyze_frame_type): New function.
	(mrk3_alloc_frame_cache): Delete function.
	(mrk3_frame_cache): Use 'mrk3_analyze_prologue' and
	'mrk3_analyze_frame_type', then finish initialisig the cache based
	on actual frame state that can't be accessed in
	'mrk3_analyze_prologue'.
	(mrk3_frame_base_address): Update for structure name changes.
	(mrk3_frame_prev_register): Remove special handling of PC.