Add generation of Secure PLT Code Signature Metadata
This is placed into a new section .debug_codesig.plt, and is the .plt
equivalent of .debug_codesig as generated by the compiler. If a secure
function is called via a secure PLT entry, this new metadata makes sure
that the PLT entry is secured by the same key as the function proper.

This reduces the number of keys that need to be manually traced throughout
the code signature system, without having to resort to having one PLT
key for the entire PLT instruction stream.

bfd/ChangeLog:

	* elf64-mrk3.c (SPLT_CODESIG_SIZE): Defined.
	(mrk3_elf_create_plt_section): Create PLT codesig metadata section.
	(mrk3_elf_check_relocs): Track size of PLT codesig metadata.
	(mrk3_elf_size_dynamic_sections): Create contents of codesig section.
	(mrk3_elf_finish_dynamic_sections): Fill contents of codesig section.