arc: Fill in the dst_mask field for the howto structures.
Currently the dst_mask field is filled in with 0 for all arc
relocations.

This makes sense, a large number of arc relocations are non-standard,
either computing the relocation value, or how the value is patched in,
is more complex than can be expressed within the howto structure.  This
is fine though, for ELF only links (which is all the arc-elf target
supports) the final relocation resolution is all done within the
arc_do_relocation function, which is called from
elf_arc_relocate_section, and this makes no use of the dst_mask field.

However, currently, the generic relocation mechanism, the one accessed
through bfd_perform_relocation has a problem.  The generic mechanism
is used in a few corner cases, when the ELF specialisation can't be
used, for example, when using objdump to dump the dwarf from an object
file, in this case the relocations within the dwarf are patched using
the generic relocation mechanism.

The generic relocation mechanism can (currently) not be taught to patch
all arc relocations, some are just too complex.  The way other targets
deal with this problem, for example xtensa, is to make use of the
special function within the howto structure to completely handle the
patching of the relocations.

The problem with this use of the special function is that some of the
relocations, for example those relating to GOT and PLT are hard to patch
correctly (if at all?) when not performing ELF final (or partial) links.

The conclusion that I reached is that overloading the special function
would not be worth the effort, we would not gain the ability to
correctly patch all relocations (the GOT/PLT relocations would not be
handled), so I instead added the dst_mask field to the howto table.

However, given that the value computed for some relocations (GOT/PLT for
example) will not be correct, filling in the dst_mask for these
relocations seems pretty pointless, better to leave the dst_mask as 0,
so it's clear that the relocation has not been patched.

And so, in this patch I have extended the relocation definitions in
arc-reloc.def, to include a dst_mask field.  For most relocations this
is 0, however, for those relocations that are likely to be used within
the dwarf I've filled in the dst_mask field with an appropriate value.

Before this commit, viewing the dwarf within an object file (using,
for example, objdump) would display 0 in all (or most) address fields.
After this commit the address fields are patched to more appropriate
values.

bfd/ChangeLog:

	* elf32-arc.c: Add dst_mask parameter to each instance of
	ARC_RELOC_HOWTO.  When creating the elf_arc_howto_table make use
	of the dst_mask parameter.

include/elf/ChangeLog:

	* arc-reloc.def: Add dst_mask attribute to all relocations.
	* arc.h: Add dst_mask parameter to ARC_RELOC_HOWTO definition.