bfd/arc: Merge ELF flags when combining bfd objects.
These changes mean that when building an object from ARC700 ELF files,
the result is an ARC700 ELF file, rather than an ARCv2 ELF file.  This
is achieved by selecting the lowest common ARC architecture as the
default, and rolling forward the machine type and flags as bfd objects
are merged in.  The result is that when we're finished we'll be at the
highest user architecture and machine type.

Previously, we started with ARCv2, and never moved back from there.

bfd/Change Log:

	* cpu-arc.c (bfd_arc_arch): Make ARC600 the default.
	* elf32-arc.c (arc_elf_merge_private_bfd_data): New function.
	(bfd_elf32_bfd_merge_private_bfd_data): Define.