bfd/arc: Allow non-instruction relocations within .text sections
On a little endian arc, a 4-byte instruction ABCD, where A is the most
significant byte, and D is the least significant byte would be stored in
memory (low to high) as BADC.  That is, each 2-byte chunk is stored in
little endian order in memory.

Currently, when a relocation is applied to such a 4-byte instruction, we
make use of arc_bfd_get_32 and arc_bfd_put_32, which perform byte
manipulation to correct for this in memory byte ordering.

This byte ordering correction is applied to all relocations within
executable sections (for little endian arc) when really the correction
should only be applied to instruction relocations; it is the instruction
fetch mechanism that loads 4-byte instructions 2-bytes at a time.

It is possible to place data into an executable section, for example it
might be more efficient to place small jump tables inline within the
code rather than placing them into a data section.

The problem then, is that these two aspects, the byte order correction,
and inline data, conflict.  Placing a 4-byte label relocation inline in
the code results in the byte-order correction being applied to it,
which, when the label is loaded using a standard arc load instruction,
returns a corrupted address.

Before the recent arc rewrite, placing data inline into the executable
sections did work.  This was thanks to using a different code path to
patch instruction related relocations, to the code that patched data
related relocations.  The instruction related relocations therefore
received the byte order correction, while the data related relocations
didn't.

After the recent rewrite this feature was lost, though I believe this
was by accident, rather than design.  This commit brings this feature
back, though the implementation is different, in order to fit with the
new arc design.

The formula field, in those relocations that should have the byte
ordering fix applied, is extended to include a call to a new macro ME,
this macro is then used within the bfd library to trigger the
application of the byte ordering fix when appropriate.  This design is
discussed here:
  https://sourceware.org/ml/binutils/2016-01/msg00021.html

bfd/ChangeLog:

	* elf32-arc.c (arc_bfd_get_32): Becomes an alias for bfd_get_32.
	(arc_bfd_put_32): Becomes an alias for bfd_put_32.
	(get_middle_endian_relocation): Delete.
	(middle_endian_convert): New function.
	(ME): Redefine, now does nothing.
	(IS_ME): New define.
	(arc_do_relocation): Extend the attached 'ARC_RELOC_HOWTO'
	definition to call middle_endian_convert.  Add a new local
	variable and make use of this throughout.

include/ChangeLog:

	* elf/arc-reloc.def: Add a call to ME within the formula for each
	relocation that requires middle-endian correction.

gas/ChangeLog:

	* testsuite/gas/arc/inline-data-1.d: New file.
	* testsuite/gas/arc/inline-data-1.s: New file.