gdb: Improve breakpoints on inline functions.
Consider the following C code example, the numbers down the left
represent line numbers:

    /* 1  */     static volatile int var;
    /* 2  */
    /* 3  */     static inline void
    /* 4  */     func (void)
    /* 5  */     {
    /* 6  */       var = 1;
    /* 7  */     }
    /* 8  */
    /* 9  */     int
    /* 10 */     main (void)
    /* 11 */     {
    /* 12 */       var = 0;
    /* 13 */       func ();
    /* 14 */       return 0;
    /* 15 */     }

The generated DWARF line table for the function main will obviously vary
depending on compiler and optimisation level, but one possible order in
which the lines could be mentioned is 11, 6, 14, the inclusion of line 6
represents the function 'func' being inlined.

When we try to place a breakpoint on line 13 using 'break file.c:13'
syntax gdb tries to find a suitable address by looking in the line table
for a mention of line 13, there is none, and so the smallest line number
after line 13 is used instead, in this case, line 14.

The user then, places a breakpoint on line 13, but stop at line 14.
This is not ideal, but could possibly be justified given the debug
information available.

However, if instead we start the test under gdb then we'll stop at line
12, after this, if we use the 'next' command gdb will step forward, and
eventually stop like this:

    (gdb) n
    13    /* 13 */       func ();

So, when we step through the code with 'next' gdb understands where line
13 appears, and knows when we've got there, yet, when we place a
breakpoint, gdb is unable to identify that same location.  This seems
even worse.

The reason that gdb is able to determine that we're on line 13 is thanks
to the block / address data that is extracted from the DWARF.  When we
stop at the PC which corresponds to the start of line 13, we look for
the smallest enclosing block, this turns out to be the block for the
inlined subroutine.  Having realised we are at the first instruction of
this block we report our location as being the call site of the inlined
function.

However, when placing a breakpoint we rely entirely on the line table to
find a suitable address.  If the line table does not include the
required information then the breakpoint will be placed at a nearby
address.

This commit tries to improve on this situation by, when parsing the
DWARF for the inlined subroutine, adding an additional entry to the line
table that maps the line of the call site to the first address of the
inline subroutine.  After this, an attempt to place a breakpoint on the
call site of an inline subroutine, should succeed.

For testing this commit I have rewritten the break-inline test.  This
test was previously not testing anything unique, it contained two tests,
one that checked we could place a temporary breakpoint on main and run
to the breakpoint, and the other test confirmed that we could issue the
"break" command with no arguments.  Both of these things are covered in
other tests, specifically break.exp.

gdb/ChangeLog:

	* buildsym.c (have_used_maybe_record_line): New global.
	(record_line_1): New function containing most of the old
	record_line, but with a new flag parameter.
	(record_line): Now a wrapper to call record_line_1 with specific
	flag.
	(maybe_record_line): New function.
	(reorder_linetable_entries): New function.
	(end_symtab_with_blockvector): Call reorder_linetable_entries.
	* buildsym.h (maybe_record_line): Declare.
	* dwarf2read.c (read_func_scope): Add information about inline
	subroutines to the line table.
	* symmisc.c (maintenance_info_line_tables): Add flag type to line
	table printing.
	* symtab.h (enum linetable_entry_flags): New enum.
	(struct linetable_entry): Add flags field.

gdb/testsuite/ChangeLog:

	* gdb.base/break-inline.c: Rewrite.
	* gdb.base/break-inline.exp: Rewrite.