mrk3/ld: Use default flags to initialise output section
When the linker script is read, the linker creates an object of type
lang_output_section_statement_type for every output section.  This
object has flags field which is initialised based on some properties
that are specified in the linker script.

When we need to examine the flags of an output section, before any input
sections are assigned to it, then we use the flags field held on the
lang_output_section_statement_type object.

When adding input sections to an output section we currently pass the
flags of the first section through to the init_os function, which is
used to create the output bfd section.  For all following sections
though, we take care of directly merging any additional flags into the
output section ourselves, outside of the init_os function.

The problem, is that any flags that are set in the
lang_output_section_statement_type object are discarded by the first
call to init_os.

The solution in this commit is to use the default flags already held
within init_os to initialise the output section, and then always take
care of manually merging the section flags outside of init_os.

This commit is a clean up rather than a bug fix.  There should be no
user visible changes after this commit.

ld/ChangeLog:

	* ldlang.c (init_os): Remove flags argument, use flags from the
	lang_output_section_statement_type instead.
	(exp_init_os): Update call to init_os.
	(lang_add_section): Likewise.
	(map_input_to_output_sections): Likewise.