ld/mrk3: Fill within a section makes the section loadable
When using fill within an output section, the section will be marked as
loadable.  The result of this is that the section contents will be
copied from the ELF into memory.

Consider the following output section description:

   .foo : {
       . += 0x10;
   } =0xab

This should create an output section of size 0x10 bytes, filled with the
byte pattern 0xab.

However, what actually happens is that the section is created, and is
the correct size, however, the section is marked as NOBITS in the ELF,
indicating that the section should be filled with zero at load time.

After this commit, using the fill will cause this section to be of type
PROGBITS, indicating the contents is held within the ELF, and should be
loaded out of the ELF at load time.

For now, using a fill of zero has the same effect, the section is marked
PROGBITS, and will have its content loaded from the ELF, even though
that content is all zero.  To keep the NOBITS behaviour, don't mention
the fill pattern at all.

ld/ChangeLog:

	* ldlang.c (insert_pad): When inserting padding, mark sections as
	loadable if they are marked allocatable.