mrk3/gdb: Use ELF address space information when attributed to pointers
This commit adds support for using DWARF DW_AT_address_class attributes to
determine whether a pointer is really 16 or 32 bits in size, even though
a compilation unit will have the pointer size set to 64 bits.

This fixes issues where without any other context, we assume that memory
we load is a valid DM32 pointer, and thus have garbage for the upper 16 bits
of a pointer. When we try to use this pointer, we end up accessing the wrong
memory.

As a side effect, this commit also adds support for the use of address space
identifiers, @DM16 for address class 1, and @DM32 for address class 2, both
in terms of output when printing types and also input when parsing types.

	* mrk3-tdep.c (MRK3_DM16_MASK): Define.
	(mrk3_p2a): Add DM16 specific address masking. Add new arg 'space'.
	(mrk3_a2p): Likewise.
	(mrk3_pointer_to_address): Determine address space, pass to mrk3_p2a.
	(mrk3_address_to_pointer): Determine address space, pass to mrk3_a2p.
	(mrk3_read_pc): Update to account for mrk3_p2a additional arg.
	(mrk3_write_pc): Update to account for mrk3_a2p additional arg.
	(mrk3_frame_cache): Likewise.
	(mrk3_unwind_pc): Likewise.
	(mrk3_unwind_stop_at_frame_p): Likewise.
	(mrk3_address_class_type_flags): Update to use DWARF address classes.
	(mrk3_address_class_type_flags_to_name): New function.
	(mrk3_address_class_name_to_type_flags): New function.
	(mrk3_gdbarch_init): Register new flag to name functions.