mrk3/ld: Merge .mrk3.records sections in output elf
The .mrk3.records sections are created by the assembler to guide the
linker when performing linker relaxation.  If we perform a partial link
then it is important that these sections are preserved so that the final
link can still perform linker relaxation correctly.  If we perform a
final link then we could discard these sections, however, keeping them
will be harmless; they don't get loaded into the final program image,
they are just meta-data that exist within the elf.

So, because of the partial linking case, in the default linker script I
will merge the .mrk3.records sections together and keep them in the
output elf file.

ld/ChangeLog:

	* scripttempl/mrk3.sc: Add entry to merge .mrk3.records sections.