mrk3/gdb: Bring gdbarch.sh and gdbarch.{c,h} into sync
The gdbarch.c and gdbarch.h files should be generated by the gdbarch.sh
script.  In the commit 98abdf295 I made a change directly to
gdbarch.{c,h}, which is not the correct way to do this.

In this commit I fix this mistake, I add a new entry to gdbarch.sh and
then regenerate gdbarch.{c,h}.  The new content is almost the same as
the content I had, and functionally it is the same; the changes are in
variable names, and code layout.

gdb/ChangeLog:

	* gdbarch.sh: Add unwind_stop_at_frame_p line.
	* gdbarch.c: Regenerate.
	* gdbarch.h: Regenerate.