gdb/mrk3: Align address before performing disassembly
The mrk3 disassembler will only disassemble from 2-byte boundaries.
This is not the expected behaviour of a gdb disassembler; usually gdb
would fetch the target memory, and we disassemble out of the memory
buffer.

For mrk3 the disassembler always disassembles from the preceeding 2-byte
boundary, with the result that gdb will think that it is disassembling
at address 0, but really you're seeing the result of disassembling at
address 0.

The resolve this issue I've added a new mechanism whereby targets can
adjust the disassembly base address as needed, in a way that is visible
to gdb.  So now, for mrk3 a request to disassemble at address 1 will
result in gdb displaying results for address 0, and knowing that this is
what it is doing.

This feature can be switched on and off with a new control flag:

  set disassemble-address-adjustment on
  set disassemble-address-adjustment off

and you can check the status of the feature with:

  show disassemble-address-adjustment

gdb/ChangeLog:

	* arch-utils.c (default_adjust_pc_for_disassembly): Define new
	function.
	* arch-utils.h (default_adjust_pc_for_disassembly): Declare new
	function.
	* disasm.c: Add 'cli/cli-cmds.h' include.
	(disassemble_address_adjustment): New global.
	(show_disassemble_address_adjustment): New function.
	(gdb_disassembly): Add call to gdbarch_adjust_pc_for_disassembly.
	(_initialize_disasm): New function.
	* gdbarch.c: Regenerate.
	* gdbarch.h: Regenerate.
	* gdbarch.sh (adjust_pc_for_disassembly): New entry.
	* mrk3-tdep.c (mrk3_adjust_pc_for_disassembly): New function.
	(mrk3_gdbarch_init): Initialise
	set_gdbarch_adjust_pc_for_disassembly.

gdb/doc/ChangeLog:

	* gdb.texinfo (Machine Code): Document set/show
	disassemble-address-adjustment commands.